refactor(backend): move validation annotation to subdirectory
