+++ backend/pom.xml
@@ -217,6 +217,8 @@
             <includes>
               <include>**/validation/*Test.java</include>
               <include>**/validation/dto/*Test.java</include>
+              <include>**/validation/annotation/dto/*Test.java</include>
+              <include>**/validation/annotation/*Test.java</include>
             </includes>
           </configuration>
         </plugin>
