+++ backend/src/main/java/com/crhistianm/springboot/gallo/springboot_gallo/validation/annotation/CorrectGenderValidator.java
@@ -1,4 +1,4 @@
-package com.crhistianm.springboot.gallo.springboot_gallo.validation;
+package com.crhistianm.springboot.gallo.springboot_gallo.validation.annotation;
 
 import org.springframework.stereotype.Component;
