+++ backend/src/test/java/com/crhistianm/springboot/gallo/springboot_gallo/validation/annotation/CorrectGenderValidatorUnitTest.java
@@ -1,4 +1,4 @@
-package com.crhistianm.springboot.gallo.springboot_gallo.validation;
+package com.crhistianm.springboot.gallo.springboot_gallo.validation.annotation;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
